cleanup of group composition plot directory
