+++ group_composition_plot/survey/Dockerfile
