+++ group_composition_plot/survey/compose.yaml
